+++ src/cli.ts
@@ -11,9 +11,8 @@ import { ZConfigReader } from './lint-config/config-reader.class';
 import { IZLintArgs } from './lint/lint-args.interface';
 import { ZLint } from './lint/lint.class';
 import { ZSassLint } from './sass-lint/sass-lint.class';
-import { ZTsLint } from './ts-lint/ts-lint.class';
-import { ZYamlLint } from './yaml-lint/yaml-lint.class';
 import { ZStyleLint } from './style-lint/style-lint.class';
+import { ZYamlLint } from './yaml-lint/yaml-lint.class';
 
 const args: IZLintArgs = usage('$0 [options]').alias('c', 'config').describe('c', 'Optional config file to use.').string('c').help().parse() as any;
 
@@ -22,7 +21,6 @@ const nullConfigReader = new ZConfigNullReader();
 const jsonConfigReader = new ZConfigReader(new ZConfigJsonParser());
 
 const htmlHint = new ZHtmlHint();
-const tsLint = new ZTsLint();
 const jsonLint = new ZJsonLint();
 const yamlLint = new ZYamlLint();
 
@@ -31,7 +29,6 @@ zLint.esLint = new ZEsLint(logger);
 zLint.sassLint = new ZSassLint(logger);
 zLint.styleLint = new ZStyleLint(logger);
 zLint.htmlHint = new ZFileLint(htmlHint, jsonConfigReader, logger, 'html');
-zLint.tsLint = new ZFileLint(tsLint, jsonConfigReader, logger, 'ts');
 zLint.jsonLint = new ZFileLint(jsonLint, nullConfigReader, logger, 'json');
 zLint.yamlLint = new ZFileLint(yamlLint, nullConfigReader, logger, 'yaml');
 zLint.run(args).then((result) => (process.exitCode = result));