+++ src/lint/lint-options.interface.ts
@@ -1,12 +1,10 @@
 export interface IZLintOptions {
   esConfig?: string;
-  tsConfig?: string;
   sassConfig?: string;
   styleConfig?: string;
   htmlConfig?: string;
 
   esFiles?: string[];
-  tsFiles?: string[];
   sassFiles?: string[];
   styleFiles?: string[];
   htmlFiles?: string[];